requirements.txt
paho-mqtt==1.6.1
watchdog==2.3.0
